+++ src/components/TransitionsModal.tsx
@@ -7,6 +7,9 @@ import { Letter, ToggleSwitchProps } from '../App.tsx';
 import ChartModal from './ChartModal.tsx';
 import ColorSwitches from './ColorSwitches.tsx';
 import BeakerModal from './BeakerModal.tsx';
+import AboutGameModal from './AboutGameModal.tsx';
+import QuestionModal from './QuestionModal.tsx';
+import CogModal from './CogModal.tsx';
 
 interface Props {
   isModalOpen: boolean;
@@ -17,7 +20,7 @@ interface Props {
   isThemeMod: boolean;
 }
 
-const customStyles = {
+export const customStyles = {
   overlay: {
     backgroundColor: 'rgba(0, 0, 0, 0.3)',
   },
@@ -33,14 +36,14 @@ const customStyles = {
   },
 };
 
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
   background-color: white;
   width: 350px;
   height: auto;
   padding: 1.2rem;
 `;
 
-const CloseButton = styled.button`
+export const CloseButton = styled.button`
   position: absolute;
   top: 0;
   right: 0;
@@ -56,7 +59,7 @@ const CloseButton = styled.button`
   padding: 0;
   margin: 0.5rem;
 `;
-const ModalText = styled.span`
+export const ModalText = styled.span`
   display: flex;
   justify-content: center;
   align-items: center;
@@ -64,37 +67,38 @@ const ModalText = styled.span`
   flex-direction: column;
   margin: 0;
 `;
-const ModalTitle = styled.span`
+export const ModalTitle = styled.span`
   margin: 0.5rem 0 1rem 0;
   font-size: 1.3rem;
 `;
-const ModalMiniText = styled.span`
+export const ModalMiniText = styled.span`
   font-size: 0.9rem;
   color: gray;
 `;
 
-const CogBox = styled.div`
+export const CogBox = styled.div`
   display: flex;
   align-items: center;
   justify-content: space-between;
   border-bottom: 1px solid gray;
 `;
 
-const CogOptionInfo = styled.div`
+export const CogOptionInfo = styled.div`
   display: flex;
   flex-direction: column;
 `;
 
-const CogOptionTitle = styled.span`
+export const CogOptionTitle = styled.span`
   font-size: 1rem;
   padding: 0.2rem 0;
 `;
 
-const CogOptionText = styled.span`
+export const CogOptionText = styled.span`
   font-size: 0.7rem;
   color: gray;
   padding: 0.2rem 0;
 `;
+
 export default function TransitionsModal({
   isModalOpen,
   handleClose,
@@ -104,79 +108,27 @@ export default function TransitionsModal({
   isThemeMod,
 }: Props) {
   const renderModalContent = () => {
-    const submitWord1: Letter[] = [
-      { letter: 'ㄱ', status: 'default' },
-      { letter: 'ㅗ', status: 'ball' },
-      { letter: 'ㅏ', status: 'ball' },
-      { letter: 'ㅈ', status: 'strike' },
-      { letter: 'ㅓ', status: 'default' },
-      { letter: 'ㅣ', status: 'ball' },
-    ];
-
     switch (modalType) {
       case 'Beaker':
         return (
           <Wrapper>
-            <ModalText>
-              <ModalTitle>개발자에게 단어 추천하기</ModalTitle>
-            </ModalText>
             <BeakerModal />
           </Wrapper>
         );
       case 'Cog':
         return (
           <Wrapper>
-            <ModalText>
-              <ModalTitle>설정</ModalTitle>
-            </ModalText>
-            <CogBox>
-              <CogOptionInfo>
-                <CogOptionTitle>사진 찍기</CogOptionTitle>
-                <CogOptionText>글자를 투명하게 만듭니다.</CogOptionText>
-              </CogOptionInfo>
-              <ColorSwitches
-                checked={isPictureMod}
-                handleSwitch={handleSwitch}
-                mode="PictureMod"
-              />
-            </CogBox>
-            <CogBox>
-              <CogOptionInfo>
-                <CogOptionTitle>고대비 형태</CogOptionTitle>
-                <CogOptionText>
-                  칸의 색깔들을 더 구별하기 쉽게 바꿉니다.
-                </CogOptionText>
-              </CogOptionInfo>
-              <ColorSwitches
-                checked={isThemeMod}
-                handleSwitch={handleSwitch}
-                mode="ThemeMod"
-              />
-            </CogBox>
+            <CogModal
+              handleSwitch={handleSwitch}
+              isPictureMod={isPictureMod}
+              isThemeMod={isThemeMod}
+            />
           </Wrapper>
         );
       case 'Question':
         return (
           <Wrapper>
-            <ModalText>
-              <ModalTitle>어떻게 할까?</ModalTitle>
-              <ModalMiniText>
-                여섯 개의 자모로 풀어쓴 한글 단어 꼬들 을 여섯 번의
-                <br />
-                도전 안에 맞혀봅시다. 한글 단어를 풀어쓴 후<br />
-                <strong>입력</strong>을 누르면 칸 색깔이 변합니다!
-              </ModalMiniText>
-              <SubmitLetterRow inputValue={submitWord1} />
-              <ModalMiniText>
-                자음 &apos;ㅈ&apos;는 올바른 자리에 있습니다. <br />
-                모음 &apos;ㅗ&apos;,&apos;ㅏ&apos;는 잘못된 자리에 있습니다.{' '}
-                <br />
-                자음 &apos;ㄱ&apos;, 모음&apos;ㅓ&apos;,&apos;ㅣ&apos;는 어느
-                곳에도 맞지 않습니다.
-                <br />
-                복합모음과 쌍자음, 겹받침은 더 작은 자모들로 풀어집니다.
-              </ModalMiniText>
-            </ModalText>
+            <QuestionModal />
           </Wrapper>
         );
       case 'Chart':
@@ -188,23 +140,7 @@ export default function TransitionsModal({
       case 'AboutGame':
         return (
           <Wrapper>
-            <ModalText>
-              <ModalTitle>이 놀이는?</ModalTitle>
-              이 놀이는 Belorin & Silas가 한국어로 만든
-              <br />
-              오픈 소스 프로젝트를 기반으로 <br />
-              구현한 게임입니다.
-              <br />
-              원본 한국어 워들 게임은
-              <a
-                href="https://kordle.kr/"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                꼬들
-              </a>
-              에서 즐기실 수 있습니다.
-            </ModalText>
+            <AboutGameModal />
           </Wrapper>
         );
       default:
